refactor: Use IProject as parameter instead of a path (LPS-119618)
Instead of using a path for generating a zip file, we are passing
a IProject object, that is not bound to the filesystem. Moreover,
we now generate and return a JSZip file instead of writing it
to a predefined path.

This way we will be allowed to transform the given projectContent
using some build tool without modifing the existing source code.