test: Avoid using arrayContaining, test for FragmentComposition values
Previously we used `expect.arrayContaining` in order to verify that
collections and fragments adehere to the correct object shape.

Unfortunately this introduces two different issues:

1. an empty array does not match — this is a problem because we *can*
   have a collection without fragments;

2. it doesn’t test for adeherence for all members of the array, it just
   verify that *at least* one member does match the expected value.

Therefore, in order to have a stricter and more realistic shape test of
the resulting ContentProject we manually iterate over `collections`,
`fragments` and `fragmentCompositions` to test their shape. In their
respective parent object are declared simply as `expect.any(Array)`.

Since the new approach does not verify the presence of members of the
array, we are not verifying that the project structure is correctly
extracted from the filesystem (could miss a fragment and we would not
notice.) In a further change we will introduce more precise testing to
verify that fragments and fragmentCompositions are actually extracted.

With this change we also now test the shape of `fragmentCompositions`,
which were previously not tested against.