test: Restore presence testing for Fragments and FragmentCompositions
In a previous change we removed the ability of tests to identify cases
where getProjectContent would miss Fragments or FragmentCompositions.

We now verify that for each collection at least something is
extracted.

While this is not optimal (we should know how many elements should be
found for each test case) this restores the previous test quality
without having to resort to manually specify for each fixtures the
expected results.